We may have fixed an issue, but... nuts, this is the current output:
ld: error in test.o(.eh_frame); no .eh_frame_hdr table will be created.

Not certain what this means, but we screwed up somewhere. The general idea of what is currently being done is, compile with the 32-bits words option for "as" and "cc". Link them with the options we always have. ld is no longer saying the library is incompatible, but now we have a strange error here...